+++ packages/alert/src/alert.tsx
@@ -6,7 +6,11 @@ import { BadgeGroup } from "@andromedaprotocol/badge";
 
 import * as React from "react";
 
-import { useAlertClass, useAlertCloseButton, useAlertIconClass } from "./styles";
+import {
+  useAlertClass,
+  useAlertCloseButton,
+  useAlertIconClass,
+} from "./styles";
 
 interface AlertContext {
   color?: string;
@@ -58,7 +62,7 @@ if (__DEV__) {
 
 interface IAlertProps extends DefaultProps {
   /* The color of the alert. */
-  color?: string;
+  color?: "gray" | "primary" | "success" | "warning" | "error" | string;
   /** Controls alert appearance */
   variant?: "subtle" | "solid" | "left-accent";
   more?:boolean;
@@ -76,41 +80,41 @@ const statuses = {
     icon: "info",
     styles: "alert-icon-info",
     label: "info",
-    new_styles : "text-white bg-gray-700 px-2.5 py-0.5",
-    new_back : "bg-gray-200",
-    new_title : "",
+    new_styles: "text-white bg-gray-700 px-2.5 py-0.5",
+    new_back: "bg-gray-200",
+    new_title: "",
   },
   primary: {
     icon: "info",
     styles: "alert-icon-info",
     label: "info",
-    new_styles : "text-white bg-primary-700 px-2.5 py-0.5",
-    new_back : "bg-primary-200 p-1",
-    new_title : "",
+    new_styles: "text-white bg-primary-700 px-2.5 py-0.5",
+    new_back: "bg-primary-200 p-1",
+    new_title: "",
   },
   success: {
     icon: "check-circle",
     styles: "alert-icon-success",
     label: "check-circle",
-    new_styles : "text-white bg-green-700 px-2.5 py-0.5",
-    new_back : "bg-green-200 p-1",
-    new_title : "",
+    new_styles: "text-white bg-green-700 px-2.5 py-0.5",
+    new_back: "bg-green-200 p-1",
+    new_title: "",
   },
   error: {
     icon: "alert-circle",
     styles: "alert-icon-error",
     label: "alert-circle",
-    new_styles : "text-white bg-pink-700 px-2.5 py-0.5",
-    new_back : "bg-pink-200",
-    new_title : "",
+    new_styles: "text-white bg-pink-700 px-2.5 py-0.5",
+    new_back: "bg-pink-200",
+    new_title: "",
   },
   warning: {
     icon: "alert-triangle",
     styles: "alert-icon-warning",
     label: "alert-triangle",
-    new_styles : "text-white bg-yellow-700 px-2.5 py-0.5",
-    new_back : "bg-yellow-200",
-    new_title : "",
+    new_styles: "text-white bg-yellow-700 px-2.5 py-0.5",
+    new_back: "bg-yellow-200",
+    new_title: "",
   },
 };
 
@@ -119,23 +123,24 @@ export const Alert = React.forwardRef<HTMLDivElement, AlertProps>(
     const { color = "primary", variant = "subtle", className, children, more,link,detail, ...rest } = props;
     const classes = useAlertClass({
       variant,
-      color
+      color,
     });
 
     const iconClasses = useAlertIconClass({
       variant,
-      color
-    })
+      color,
+    });
 
     const context = { color };
-    const { icon, styles, label, new_styles , new_back,new_title} = statuses[color] || {};
-    
+    const { icon, styles, label, new_styles, new_back, new_title } =
+      statuses[color] || {};
+
     return (
       <AlertContext.Provider value={context}>
         <div
           role="alert"
           ref={ref}
-          className={cx(classes, className,styles)}
+          className={cx(classes, className, styles)}
           {...rest}
         >
           {/* { mark && (
@@ -151,27 +156,38 @@ export const Alert = React.forwardRef<HTMLDivElement, AlertProps>(
                   <div className={cx("rounded-full flex justify-center items-center", new_styles)}>{badge}</div>
                   <div className={cx("ml-3 mr-2", new_title)}>{title}</div>
               </div>
-            )
-          }
-          { mark === 'icon' && (
-              <div className={cx("flex flex-row p-1 mr-2 rounded-full justify-center text-center",iconClasses)}>
-                  <div className="flex w-4 h-4 m-auto"><Icon name={icon} label=""/></div>
-                  <div className={cx("ml-3 mr-2", new_title)}>{title}</div>
+              <div className={cx("ml-3 mr-2", new_title)}>{title}</div>
+            </div>
+          )}
+          {mark === "icon" && (
+            <div
+              className={cx(
+                "flex flex-row p-1 mr-2 rounded-full justify-center text-center",
+                iconClasses
+              )}
+            >
+              <div className="flex w-4 h-4 m-auto">
+                <Icon name={icon} label="" />
               </div>
             )
           } */}
           <div className={cx('flex flex-row space-x-4 items-center','')}>
             {children}
           </div>
-          {
-            more && (
-              <div className="mt-3 ml-0 cursor-pointer md:m-auto md:mr-0">
-                <span className="md:ml-3" onClick={detail}>Learn more</span>
-                <span className="ml-3">View changes</span>
-                <a className="no-underline" href={(link ? link : '')}><span className="ml-1"><Icon className="w-3 h-3" name='arrow-right' label={""}/></span></a>
-              </div>
-            )
-          }
+          {/** TODO */
+          more && (
+            <div className="mt-3 ml-0 cursor-pointer md:m-auto md:mr-0">
+              <span className="md:ml-3" onClick={detail}>
+                Learn more
+              </span>
+              <span className="ml-3">View changes</span>
+              <a className="no-underline" href={link ? link : ""}>
+                <span className="ml-1">
+                  <Icon className="w-3 h-3" name="arrow-right" label={""} />
+                </span>
+              </a>
+            </div>
+          )}
         </div>
       </AlertContext.Provider>
     );